Changed ClientGroupId in BitfinexNewOrder to Long instead of Integer: https://github.com/jnidzwetzki/bitfinex-v2-wss-api-java/issues/104 (#105)

+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/BitfinexOrderBuilder.java
@@ -36,7 +36,7 @@ public class BitfinexOrderBuilder {
 	private BigDecimal priceTrailing;
 	private BigDecimal priceAuxLimit;
 	private Set<BitfinexOrderFlag> orderFlags;
-	private int groupid = -1;
+	private long groupid = -1;
 
 	private BitfinexOrderBuilder(final BitfinexCurrencyPair symbol, final BitfinexOrderType type, 
 			final BigDecimal amount) {
